Create main.c
This is the main function file for Nucleo-64 to perform the CMSIS - DSP FIR filter on the ECG noise signal. The ECG noise signal was stored in sd card, this main,c will do the function of reading the noise file, processing the pulse and writing back to the sd card.